requirements.txt
# OpenClaw-ROS-Bridge Core Dependencies
# Python 3.8+ compatible (ROS1) / 3.10+ compatible (ROS2 Jazzy)

# Core Config/Utils (Required)
pyyaml>=6.0.1,<7.0
python-dotenv>=1.0.0,<2.0
psutil>=5.9.5,<6.0
requests>=2.31.0,<3.0
pyserial>=3.5,<4.0
pyusb>=1.2.1
smbus2>=0.4.2,<1.0

# Observability/Visualization
matplotlib>=3.7.2,<4.0
seaborn>=0.12.2,<0.13.0

# Testing
pytest>=7.4.0,<9.0
pytest-cov>=4.1.0,<6.0
pytest-benchmark>=4.0.0,<5.0
mock>=5.1.0,<6.0
pytest-mock>=3.12.0,<4.0

# Documentation
sphinx>=7.2.6,<8.0
sphinx-rtd-theme>=1.3.0,<2.0
sphinx-autodoc-typehints>=1.25.0,<2.0

# ============================================
# ROS PACKAGES - Install via apt on Ubuntu
# ============================================
# rospy>=1.15.0          # sudo apt install python3-rospy
# rclpy>=3.8.0           # sudo apt install python3-rclpy
# std-msgs               # sudo apt install ros-${ROS_DISTRO}-std-msgs
# sensor-msgs            # sudo apt install ros-${ROS_DISTRO}-sensor-msgs
# ros1_bridge            # sudo apt install ros-${ROS_DISTRO}-ros1-bridge

# ============================================
# EMBEDDED PACKAGES - Platform specific
# ============================================
# jetson-stats>=4.0.0    # Jetson only
# rpi.gpio>=0.7.1        # Raspberry Pi only  
# nvidia-ml-py>=11.525   # NVIDIA GPU only
